fix(rating) - changing overall_rating to f64

+++ migrations/20240125152759_audibook-change_overall_rating_to_float.down.sql
@@ -1,1 +1,1 @@
-ALTER TABLE "Audiobook" ALTER COLUMN overall_rating type decimal(2,2);
+ALTER TABLE "Audiobook" ALTER COLUMN overall_rating type smallint;
