+++ migrations/20240125152759_audibook-change_overall_rating_to_float.up.sql
@@ -1,1 +1,2 @@
-ALTER TABLE "Audiobook" ALTER COLUMN overall_rating type smallint;
+
+ALTER TABLE "Audiobook" ALTER COLUMN overall_rating type float8;
